+++ extensions/algebra/circuit/src/extension/hybrid.rs
@@ -31,6 +31,7 @@ use openvm_riscv_circuit::{adapters::U16_BITS, Rv64ImGpuProverExt};
 use openvm_stark_backend::prover::AirProvingContext;
 
 use crate::{
+    cuda::ModularIsEqualChipGpu,
     fp2_chip::{get_fp2_addsub_chip, get_fp2_muldiv_chip, Fp2Air, Fp2Chip},
     modular_chip::*,
     AlgebraRecord, Fp2Extension, ModularExtension, Rv64ModularConfig, Rv64ModularWithFp2Config,
@@ -435,145 +436,5 @@ impl VmBuilder<E> for Rv64ModularWithFp2HybridBuilder {
             inventory,
         )?;
         Ok(chip_complex)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// GPU tracegen for ModularIsEqual (dedicated kernel; see cuda/src/modular_is_eq.cu)
-// ---------------------------------------------------------------------------
-
-mod is_eq_cuda_abi {
-    #![allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
-    use openvm_cuda_backend::prelude::F;
-    use openvm_cuda_common::{d_buffer::DeviceBuffer, error::CudaError, stream::cudaStream_t};
-
-    macro_rules! declare_is_eq_launcher {
-        ($name:ident) => {
-            extern "C" {
-                fn $name(
-                    d_trace: *mut F,
-                    height: usize,
-                    rows_used: usize,
-                    d_records: *const u8,
-                    rec_stride: usize,
-                    rec_core_offset: usize,
-                    d_modulus_limbs: *const u16,
-                    d_range_checker: *mut u32,
-                    rc_bins: usize,
-                    pointer_max_bits: u32,
-                    timestamp_max_bits: u32,
-                    stream: cudaStream_t,
-                ) -> i32;
-            }
-        };
-    }
-    declare_is_eq_launcher!(_modular_is_eq_tracegen_l4);
-    declare_is_eq_launcher!(_modular_is_eq_tracegen_l6);
-
-    pub unsafe fn tracegen(
-        d_trace: &DeviceBuffer<F>,
-        height: usize,
-        rows_used: usize,
-        d_records: &DeviceBuffer<u8>,
-        rec_stride: usize,
-        rec_core_offset: usize,
-        d_modulus_limbs: &DeviceBuffer<u16>,
-        d_range_checker: &DeviceBuffer<F>,
-        num_lanes: usize,
-        pointer_max_bits: u32,
-        timestamp_max_bits: u32,
-        stream: cudaStream_t,
-    ) -> Result<(), CudaError> {
-        let launcher = match num_lanes {
-            4 => _modular_is_eq_tracegen_l4,
-            6 => _modular_is_eq_tracegen_l6,
-            _ => panic!("unsupported ModularIsEqual num_lanes {num_lanes}"),
-        };
-        CudaError::from_result(launcher(
-            d_trace.as_mut_ptr(),
-            height,
-            rows_used,
-            d_records.as_ptr(),
-            rec_stride,
-            rec_core_offset,
-            d_modulus_limbs.as_ptr(),
-            d_range_checker.as_mut_ptr() as *mut u32,
-            d_range_checker.len(),
-            pointer_max_bits,
-            timestamp_max_bits,
-            stream,
-        ))
-    }
-}
-
-pub struct ModularIsEqualChipGpu<const NUM_LANES: usize, const TOTAL_LIMBS: usize> {
-    range_checker: Arc<VariableRangeCheckerChipGPU>,
-    d_modulus: openvm_cuda_common::d_buffer::DeviceBuffer<u16>,
-    pointer_max_bits: u32,
-    timestamp_max_bits: u32,
-}
-
-impl<const NUM_LANES: usize, const TOTAL_LIMBS: usize>
-    ModularIsEqualChipGpu<NUM_LANES, TOTAL_LIMBS>
-{
-    pub fn new(
-        modulus_limbs: [u16; TOTAL_LIMBS],
-        byte_ptr_max_bits: usize,
-        timestamp_max_bits: usize,
-        range_checker: Arc<VariableRangeCheckerChipGPU>,
-    ) -> Self {
-        use openvm_cuda_common::copy::MemCopyH2D;
-        let d_modulus = modulus_limbs
-            .as_slice()
-            .to_device_on(&range_checker.device_ctx)
-            .unwrap();
-        Self {
-            range_checker,
-            d_modulus,
-            pointer_max_bits: byte_ptr_max_bits as u32,
-            timestamp_max_bits: timestamp_max_bits as u32,
-        }
-    }
-}
-
-impl<const NUM_LANES: usize, const TOTAL_LIMBS: usize> Chip<DenseRecordArena, GpuBackend>
-    for ModularIsEqualChipGpu<NUM_LANES, TOTAL_LIMBS>
-{
-    fn generate_proving_ctx(&self, arena: DenseRecordArena) -> AirProvingContext<GpuBackend> {
-        use openvm_cuda_common::copy::MemCopyH2D;
-        type ARec<const L: usize> = Rv64IsEqualModU16AdapterRecord<2, L>;
-        let record_stride = size_of::<(ARec<NUM_LANES>, ModularIsEqualRecord<TOTAL_LIMBS>)>();
-        let rec_core_offset = size_of::<ARec<NUM_LANES>>();
-        let records = arena.allocated();
-        if records.is_empty() {
-            return AirProvingContext::simple_no_pis(DeviceMatrix::dummy());
-        }
-        debug_assert_eq!(records.len() % record_stride, 0);
-        let rows_used = records.len() / record_stride;
-        let height = rows_used.next_power_of_two();
-        let width = Rv64IsEqualModU16AdapterCols::<F, 2, NUM_LANES>::width()
-            + ModularIsEqualCoreCols::<F, TOTAL_LIMBS>::width();
-
-        let device_ctx = &self.range_checker.device_ctx;
-        let d_records = records.to_device_on(device_ctx).unwrap();
-        let d_trace = DeviceMatrix::<F>::with_capacity_on(height, width, device_ctx);
-        unsafe {
-            is_eq_cuda_abi::tracegen(
-                d_trace.buffer(),
-                height,
-                rows_used,
-                &d_records,
-                record_stride,
-                rec_core_offset,
-                &self.d_modulus,
-                &self.range_checker.count,
-                NUM_LANES,
-                self.pointer_max_bits,
-                self.timestamp_max_bits,
-                device_ctx.stream.as_raw(),
-            )
-            .unwrap();
-        }
-        AirProvingContext::simple_no_pis(d_trace)
     }
 }
